refactor(progress): type flush interval constant as uint32_t
static constexpr unsigned long -> uint32_t for FLUSH_INTERVAL_MS (120000,
compile-time only, no RAM footprint). Seconds-since-boot timestamp
lastFlushSec_ stays uint32_t alongside it.

+++ src/ProgressManager.h
@@ -33,7 +33,7 @@
 // stall can only block the flusher, never the render task.
 class ProgressManager {
  public:
-  static constexpr unsigned long FLUSH_INTERVAL_MS = 120000;
+  static constexpr uint32_t FLUSH_INTERVAL_MS = 120000;
   // On-disk record: 6 bytes of spine/page/count (+4 bytes visibleTextOffset
   // when known). Byte order matches the firmware's historical layout — do
   // not reorder without a load-side migration. The offset participates in
